moved triage unit test to its own file
fixed triage unit test
